+++ node-src/graphQL/Book/schema.js
@@ -18,7 +18,7 @@ export const type = `
     dateAdded: String
     editorialReviews: [EditorialReview]
     similarItems: [String]
-    timestamp: Int
+    timestamp: Float
     similarBooks(SORT: BookSummarySort, SORTS: [BookSummarySort]): [BookSummary]
   }
 
@@ -63,7 +63,7 @@ export const type = `
     dateAdded: String
     editorialReviews: [EditorialReviewInput]
     similarItems: [String]
-    timestamp: Int
+    timestamp: Float
     similarBooks: [BookSummaryInput]
   }
 
@@ -115,7 +115,7 @@ export const type = `
     similarItems_UPDATES: [StringArrayUpdate]
     similarItems_PULL: [String]
     similarItems_ADDTOSET: [String]
-    timestamp: Int
+    timestamp: Float
     timestamp_INC: Int
     timestamp_DEC: Int
     similarBooks_ADD: [BookSummaryInput]
@@ -262,13 +262,13 @@ export const type = `
     similarItems_contains: String
     similarItems_containsAny: [String]
     similarItems_ne: [String]
-    timestamp_lt: Int
-    timestamp_lte: Int
-    timestamp_gt: Int
-    timestamp_gte: Int
-    timestamp: Int
-    timestamp_ne: Int
-    timestamp_in: [Int]
+    timestamp_lt: Float
+    timestamp_lte: Float
+    timestamp_gt: Float
+    timestamp_gte: Float
+    timestamp: Float
+    timestamp_ne: Float
+    timestamp_in: [Float]
     OR: [BookFilters]
   }
   
@@ -425,13 +425,13 @@ export const query = `
     similarItems_contains: String,
     similarItems_containsAny: [String],
     similarItems_ne: [String],
-    timestamp_lt: Int,
-    timestamp_lte: Int,
-    timestamp_gt: Int,
-    timestamp_gte: Int,
-    timestamp: Int,
-    timestamp_ne: Int,
-    timestamp_in: [Int],
+    timestamp_lt: Float,
+    timestamp_lte: Float,
+    timestamp_gt: Float,
+    timestamp_gte: Float,
+    timestamp: Float,
+    timestamp_ne: Float,
+    timestamp_in: [Float],
     OR: [BookFilters],
     SORT: BookSort,
     SORTS: [BookSort],
